+++ shared-engine/src/commonMain/kotlin/com/example/timeboxvibe/engine/core/MainMenuScene.kt
@@ -7,6 +7,5 @@ object MainMenuScene : Scene {
     override fun update(dt: Float) {}
     override fun render(renderer: ScaledProceduralRenderer, playX: Int, playY: Int, playW: Int, playH: Int) {}
     override fun onInput(inputCode: Int) {}
-    override fun onTouch(x: Int, y: Int, action: Int, playX: Int, playY: Int, playW: Int, playH: Int) {}
+    override fun onTouch(x: Float, y: Float, action: Int, playX: Int, playY: Int, playW: Int, playH: Int) {}
 }
-
