+++ shared-engine/src/commonMain/kotlin/com/example/timeboxvibe/engine/core/Scene.kt
@@ -14,6 +14,6 @@ interface Scene {
     fun render(renderer: ScaledProceduralRenderer, playX: Int, playY: Int, playW: Int, playH: Int)
     fun renderOverlay(renderer: ScaledProceduralRenderer, playX: Int, playY: Int, playW: Int, playH: Int) {}
     fun onInput(inputCode: Int)
-    fun onTouch(x: Int, y: Int, action: Int, playX: Int, playY: Int, playW: Int, playH: Int)
+    fun onTouch(x: Float, y: Float, action: Int, playX: Int, playY: Int, playW: Int, playH: Int)
     fun onInput(x: Int, y: Int, action: Int, playX: Int, playY: Int, playW: Int, playH: Int) {}
 }
