+++ shared-engine/src/commonMain/kotlin/com/example/timeboxvibe/engine/core/TemplateForgeScene.kt
@@ -213,7 +213,7 @@ object TemplateForgeScene : Scene {
         }
     }
 
-    override fun onTouch(x: Int, y: Int, action: Int, playX: Int, playY: Int, playW: Int, playH: Int) {
+    override fun onTouch(x: Float, y: Float, action: Int, playX: Int, playY: Int, playW: Int, playH: Int) {
         val playAreaH = playH.toFloat()
         val logicalHeight = SceneManager.logicalHeight
         val inPlayArea = x >= playX && y >= playY && x < playX + playW && y < playY + playH
@@ -246,7 +246,7 @@ object TemplateForgeScene : Scene {
                     val deltaX = x - initialTouchX
                     val deltaY = y - initialTouchY
                     if (inPlayArea && abs(deltaX) < (U / 2).toFloat() && abs(deltaY) < (U / 2).toFloat() && !hasDragged) {
-                        onInput(x, y, TouchAction.UP, playX, playY, playW, playH)
+                        onInput(x.toInt(), y.toInt(), TouchAction.UP, playX, playY, playW, playH)
                     }
                 }
             }
